Added Lorenz' patch set for the math module, addressing:   * consistently use degrees and not radians   * raise a ValueError rather than a ZeroDivisionError in case the div     zero occurs during an internal calculation.   * SLERP implementation is more robust. check for special cases like     angle=0,180,360 degrees.   * fixed a bug in LERP. t must be in range [0,1]   * in the rotation methods: check if the rotation axis is zero and raise     a ValueError in case it is.   * check for div zero for div, idiv, floor_div   * also check for div zero in the angleto method   * a wrong function signature in basemod.c

+++ src/base/basemod.c
@@ -108,8 +108,8 @@ LongFromObj (PyObject* obj, long* val)
     return 0;
 }
 
-unsigned long
-UlongFromObj (PyObject* obj, long* val)
+int
+UlongFromObj (PyObject* obj, unsigned long* val)
 {
     PyObject* longobj;
     unsigned long tmp;
